# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea EVA3",
        "description": "API para gestión de candidatos con CV ciego, empresas, relación empresa-candidato y salud de la API.\n\n**Rate Limiting:** 60 requests/minuto por IP.\n**Caché:** Endpoints GET cacheados por 5 minutos.\n**Tiempo de respuesta objetivo:** < 200ms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local Docker"
        }
    ],
    "paths": {
        "/candidatos": {
            "get": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Listar candidatos",
                "description": "Obtiene todos los candidatos usando CV ciego. No retorna nombre, email, teléfono, edad, género ni comuna.",
                "operationId": "0ec810ccabcaed118fdedde772f1255d",
                "responses": {
                    "200": {
                        "description": "Lista de candidatos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CandidatoPublico"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Crear candidato",
                "description": "Crea un candidato guardando datos privados, pero las respuestas mantienen el formato de CV ciego.",
                "operationId": "a9e458e368ac996c35d95125efe8bc44",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CandidatoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidato creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CandidatoPublico"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/candidatos/{id}": {
            "get": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Obtener candidato por ID",
                "description": "Obtiene un candidato específico mediante CV ciego.",
                "operationId": "d962389ca35587c72c9f7ea3dcef16ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del candidato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidato encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CandidatoPublico"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidato no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Actualizar candidato",
                "description": "Actualiza los datos de un candidato. La respuesta mantiene el CV ciego.",
                "operationId": "0aecc4523bf38b0ae4c15e94dbf4ae67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del candidato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CandidatoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidato actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CandidatoPublico"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidato no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Eliminar candidato",
                "description": "Elimina un candidato por ID.",
                "operationId": "1d85752cbc063a65a60a0ef5026f4ee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del candidato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidato eliminado correctamente"
                    },
                    "404": {
                        "description": "Candidato no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/candidatos/{id}/estado": {
            "patch": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Activar o desactivar candidato",
                "description": "Permite encender o apagar el estado del candidato.",
                "operationId": "ed1ed12396eb482ba308e25b6c309789",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del candidato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CandidatoPublico"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidato no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CandidatoPublico": {
                "title": "Candidato público",
                "description": "Respuesta pública del candidato usando CV ciego. No incluye nombre, email, teléfono, edad, género ni comuna.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "example": "Técnico Nivel Superior"
                    },
                    "experiencia_laboral": {
                        "type": "string",
                        "example": "3 años en soporte técnico"
                    },
                    "habilidades": {
                        "type": "string",
                        "example": "PHP, Laravel, MySQL"
                    },
                    "certificaciones": {
                        "type": "string",
                        "example": "Certificación Desarrollo Backend"
                    },
                    "referencias": {
                        "type": "string",
                        "example": "Referencia laboral disponible"
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "CandidatoRequest": {
                "title": "Crear candidato",
                "required": [
                    "nombre",
                    "email",
                    "nivel_educacional",
                    "experiencia_laboral",
                    "habilidades"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan@test.cl"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56911111111"
                    },
                    "edad": {
                        "type": "integer",
                        "example": 28
                    },
                    "genero": {
                        "type": "string",
                        "example": "Masculino"
                    },
                    "comuna_residencia": {
                        "type": "string",
                        "example": "Providencia"
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "example": "Universitario completo"
                    },
                    "experiencia_laboral": {
                        "type": "string",
                        "example": "3 años como desarrollador backend"
                    },
                    "habilidades": {
                        "type": "string",
                        "example": "PHP, Laravel, MySQL"
                    },
                    "certificaciones": {
                        "type": "string",
                        "example": "AWS Practitioner"
                    },
                    "referencias": {
                        "type": "string",
                        "example": "Empresa ABC"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}